Refactor: ASファイルネームとrouteを変更
